✅ basic cart-Utils : calculateCartTotal
