+++ src/refactoring/models/cart.ts
@@ -1,4 +1,4 @@
-import { CartItem, Coupon } from "../../types";
+import { CartItem, Coupon } from '../../types';
 
 export const calculateItemTotal = (item: CartItem) => {
   return 0;
@@ -8,21 +8,58 @@ export const getMaxApplicableDiscount = (item: CartItem) => {
   return 0;
 };
 
+export type TotalPrices = {
+  totalBeforeDiscount: number;
+  totalAfterDiscount: number;
+  totalDiscount: number;
+};
+
 export const calculateCartTotal = (
   cart: CartItem[],
-  selectedCoupon: Coupon | null
-) => {
+  selectedCoupon: Coupon | null,
+): TotalPrices => {
+  let totalBeforeDiscount = 0;
+  let totalAfterDiscount = 0;
+
+  cart.forEach((item) => {
+    const { price } = item.product;
+
+    const { quantity } = item;
+    totalBeforeDiscount += price * quantity;
+
+    const discount = item.product.discounts.reduce((maxDiscount, d) => {
+      return quantity >= d.quantity && d.rate > maxDiscount
+        ? d.rate
+        : maxDiscount;
+    }, 0);
+    totalAfterDiscount += price * quantity * (1 - discount);
+  });
+
+  let totalDiscount = totalBeforeDiscount - totalAfterDiscount;
+
+  if (selectedCoupon) {
+    if (selectedCoupon.discountType === 'amount') {
+      totalAfterDiscount = Math.max(
+        0,
+        totalAfterDiscount - selectedCoupon.discountValue,
+      );
+    } else {
+      totalAfterDiscount *= 1 - selectedCoupon.discountValue / 100;
+    }
+    totalDiscount = totalBeforeDiscount - totalAfterDiscount;
+  }
+
   return {
-    totalBeforeDiscount: 0,
-    totalAfterDiscount: 0,
-    totalDiscount: 0,
+    totalBeforeDiscount: Math.round(totalBeforeDiscount),
+    totalAfterDiscount: Math.round(totalAfterDiscount),
+    totalDiscount: Math.round(totalDiscount),
   };
 };
 
 export const updateCartItemQuantity = (
   cart: CartItem[],
   productId: string,
-  newQuantity: number
+  newQuantity: number,
 ): CartItem[] => {
   return [];
 };
